Rollup lots of minor issues
This includes changes from #11 and #12 along with lots of minor things.  For example, it uses `execFile*` instead of `exec*` which addresses issues with spaces in file names on Windows.  BUT, by using `execFile*`, we need to use `yarn.cmd` on Windows when running `yarn`, so it handles that.  Also, at least when I use the `tar` that comes with `git`/`bash` on Windows, it can handle `\\` in file names, but it doesn't like the `C:\\` at the start.  So I changed the `tar` command to use relative path names.

These are all the issues I could find so far that were causing me grief on Windows so this addresses all the issues I raised in #8.